Split rules into multiple files
